requirements.txt
# AssureXRay Dependencies
# Python >=3.10, <3.13 (3.12 recommended)
#
# For GPU (CUDA 12.1):
#   pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu121
#
# For CPU only (Mac/development):
#   pip install -r requirements.txt

# Core ML framework (must be a matched pair)
torch==2.6.0
torchvision==0.21.0

# Model zoo
timm==1.0.15

# Augmentation (MIT-licensed archived version -- stable, no active maintenance)
albumentations==2.0.8

# Explainability
grad-cam==1.5.5

# Evaluation
scikit-learn==1.7.2
scipy==1.15.2

# Data handling
pandas==2.2.3
numpy==2.2.3
Pillow==11.1.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2

# Utilities
tqdm==4.67.1
pyyaml==6.0.2

# Download and audit
requests>=2.31.0
imagehash==4.3.2
openpyxl>=3.1.0

# Demo
streamlit>=1.42.0
